Refer to description below
Added Store and Update function for:
-Admission
-BMVS
-Dsicharge

All 3 store functions have been tested and should be working correctly but would need to re-test tomorrow to be sure.

Update function for all 3 forms can't be tested yet since the data for SP2-4 is still unable to be passed into the views to process old data.
(This should be done by tomorrow in the edit function of the Admssion_Controller)