Refer to description below
The Follow Up Quesitionnaire for the moment is suddenly not working. Do not try to alter anything regarding it as I am trying figure out what happened.

Changed the route for Follow Up questionnaire and placed it accordingly.

Changed the button layout a little in the studyspecific page.

Added if function in Admission Controller for Conclusion of participation.

Chanfged namer of the update function in PostStudy controllerto updateFollowUpQ as it is only used to update Follow Up Questionnaire.